LPD-76015 Reorganizing files in the components folder
