+++ modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/js/components/InfoSummary.tsx
