+++ modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/js/components/props_transformer/ProjectsFDSPropsTransformer.tsx
@@ -15,7 +15,7 @@ import {
 } from '@liferay/site-cms-site-initializer';
 import {fetch} from 'frontend-js-web';
 
-import StateLabel from '../../common/components/StateLabel';
+import StateLabel from '../StateLabel';
 import UserRelationshipRenderer from './cell_renderers/UserRelationshipRenderer';
 
 type Action = {
