+++ modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/js/index.ts
@@ -3,8 +3,8 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {default as EditorToolbar} from './editor/components/EditorToolbar';
-export {default as ProjectInfoSummary} from './main_view/projects/components/ProjectInfoSummary';
-export {default as ProjectsFDSPropsTransformer} from './main_view/props_transformer/ProjectsFDSPropsTransformer';
-export {default as TasksFDSPropsTransformer} from './main_view/props_transformer/TasksFDSPropsTransformer';
-export {default as TaskInfoSummary} from './main_view/tasks/components/TaskInfoSummary';
+export {default as EditorToolbar} from './components/EditorToolbar';
+export {default as ProjectInfoSummary} from './components/project/ProjectInfoSummary';
+export {default as ProjectsFDSPropsTransformer} from './components/props_transformer/ProjectsFDSPropsTransformer';
+export {default as TasksFDSPropsTransformer} from './components/props_transformer/TasksFDSPropsTransformer';
+export {default as TaskInfoSummary} from './components/task/TaskInfoSummary';
