--- modules/apps/site/site-cmp-site-initializer/src/main/resources/META-INF/resources/js/main_view/tasks/components/TaskInfoSummary.tsx
@@ -1,94 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import Label from '@clayui/label';
-import {
-	Assignee,
-	AssigneeValue,
-} from '@liferay/object-dynamic-data-mapping-form-field-type';
-import React from 'react';
-
-import InfoSummary from '../../../common/components/InfoSummary';
-import StateSelector, {State} from '../../../common/components/StateSelector';
-import {patchTaskById} from '../../../utils/api';
-import {DISPLAY_TYPES} from '../../../utils/constants';
-
-interface TaskInfoSummaryProps {
-	assignTo: AssigneeValue;
-	dueDate: string;
-	initialState: string;
-	states: State[];
-	tags: string[];
-	taskId: string;
-}
-
-export default function TaskInfoSummary({
-	assignTo,
-	dueDate,
-	initialState,
-	states,
-	tags,
-	taskId,
-}: TaskInfoSummaryProps) {
-	return (
-		<InfoSummary
-			defaultOpen={true}
-			items={[
-				{
-					label: 'State',
-					value: (
-						<StateSelector
-							initialSelectedKey={initialState}
-							onChange={async (key: string) => {
-								await patchTaskById({
-									body: {state: key},
-									taskId,
-								});
-							}}
-							states={states}
-						/>
-					),
-				},
-				{
-					label: 'Assignee',
-					value: (
-						<Assignee
-							name="assignee"
-							onChange={async (value: any) => {
-								await patchTaskById({
-									body: {assignTo: value},
-									taskId,
-								});
-							}}
-							searchURL={
-								Liferay.ThemeDisplay.getPortalURL() +
-								'/o/cmp/assignee-context/'
-							}
-							showLabel={false}
-							value={assignTo}
-							visible={true}
-						/>
-					),
-				},
-				{label: 'Due Date', value: dueDate},
-				{
-					label: 'Tags',
-					value: (
-						<div>
-							{tags.map((tag, index) => (
-								<Label
-									displayType={DISPLAY_TYPES[index % 6]}
-									key={tag}
-								>
-									{tag}
-								</Label>
-							))}
-						</div>
-					),
-				},
-			]}
-		/>
-	);
-}
